refactor(home): verticals css, adds warning icons
This commit refactors the vertical lines css on the home page to be
anchored from the center of the screen instead of the left or right.
This improves the visuals on wide screens and also works better on
smaller screens.

The commit also adds two warning icons that surround the work in
progress header to highlight this even more.